asdasdqweqweasd

+++ ft_u.c
@@ -62,9 +62,9 @@ char    *ft_itoa_basel(unsigned long long int n, int b)
 
 int ft_o(t_pf *f, va_list pa)
 {
-	unsigned int i;
+	unsigned long long int i;
 
-	i = va_arg(pa, unsigned int);
+	i = va_arg(pa, unsigned long long int);
 	if (f->t.ht == 1 && i != 0)
 		f->t.res += ft_putchar('0');
 	f->t.nb = ft_itoa_basel(i, 8);
